refactor(alertmanager): remove dead legacy config and tidy integration test helpers
